Rename full_uncertainty.py to spall_uncertainty.py and update imports
