+++ src/alpss/__init__.py
@@ -11,7 +11,7 @@ from alpss.carrier.frequency import carrier_frequency
 from alpss.carrier.filter import carrier_filter
 from alpss.velocity.calculation import velocity_calculation
 from alpss.analysis.spall import spall_analysis
-from alpss.analysis.full_uncertainty import spall_uncertainty_analysis
+from alpss.analysis.spall_uncertainty import spall_uncertainty_analysis
 from alpss.analysis.instantaneous_uncertainty import instantaneous_uncertainty_analysis
 from alpss.plotting.plots import plot_results, plot_voltage
 from alpss.io.saving import save
