+++ src/alpss/analysis/__init__.py
@@ -1,4 +1,4 @@
 from alpss.analysis.spall import spall_analysis
 from alpss.analysis.instantaneous_uncertainty import instantaneous_uncertainty_analysis
-from alpss.analysis.full_uncertainty import spall_uncertainty_analysis
+from alpss.analysis.spall_uncertainty import spall_uncertainty_analysis
 from alpss.analysis.hel import hel_detection, elastic_shock_strain_rate, HELResult
