+++ src/alpss/utils/phases.py
@@ -13,7 +13,7 @@ from alpss.velocity.calculation import velocity_calculation
 from alpss.analysis.instantaneous_uncertainty import instantaneous_uncertainty_analysis
 from alpss.analysis.velocity_uncertainty import velocity_uncertainty_analysis
 from alpss.analysis.spall import spall_analysis
-from alpss.analysis.full_uncertainty import spall_uncertainty_analysis
+from alpss.analysis.spall_uncertainty import spall_uncertainty_analysis
 from alpss.analysis.shock import shock_analysis
 from alpss.analysis.hel import hel_detection
 from alpss.plotting.plots import plot_results
